+++ apps/contribute/src/screens/Uploader/UploadProgress/index.tsx
@@ -1,23 +1,22 @@
 import { Link } from "@phylopic/api-models"
-import { getImageFileExtension, Hash, ImageMediaType, isHash } from "@phylopic/utils"
+import { createSearch, getImageFileExtension, Hash, ImageMediaType, isHash, UUID } from "@phylopic/utils"
 import axios, { AxiosProgressEvent } from "axios"
 import { FC, useEffect, useMemo, useState } from "react"
 import useAuthToken from "~/auth/hooks/useAuthToken"
 import useContributorUUID from "~/profile/useContributorUUID"
 import Dialogue from "~/ui/Dialogue"
-import { ICON_ARROW_LEFT } from "~/ui/ICON_SYMBOLS"
 import Speech from "~/ui/Speech"
-import UserButton from "~/ui/UserButton"
-import UserOptions from "~/ui/UserOptions"
 import styles from "./index.module.scss"
+import UploadError from "./UploadError"
 export interface Props {
     buffer: Buffer
     filename?: string
     onCancel: () => void
     onComplete: (hash: Hash) => void
+    existingUUID?: UUID
     type: ImageMediaType
 }
-const UploadProgress: FC<Props> = ({ buffer, filename, onCancel, onComplete, type }) => {
+const UploadProgress: FC<Props> = ({ buffer, filename, onCancel, onComplete, existingUUID, type }) => {
     const token = useAuthToken()
     const contributorUUID = useContributorUUID()
     const [loaded, setLoaded] = useState(0)
@@ -30,18 +29,22 @@ const UploadProgress: FC<Props> = ({ buffer, filename, onCancel, onComplete, typ
     useEffect(() => {
         if (buffer && contributorUUID && token) {
             const controller = new AbortController()
-            const promise = axios.post<Link>(`${process.env.NEXT_PUBLIC_API_URL}/uploads`, buffer, {
-                headers: {
-                    authorization: `Bearer ${token}`,
-                    "content-type": type,
+            const promise = axios.post<Link>(
+                `${process.env.NEXT_PUBLIC_API_URL}/uploads${createSearch({ existing_uuid: existingUUID })}`,
+                buffer,
+                {
+                    headers: {
+                        authorization: `Bearer ${token}`,
+                        "content-type": type,
+                    },
+                    onUploadProgress: (event: AxiosProgressEvent) => {
+                        setLoaded(event.loaded)
+                        setTotal(event.total ?? NaN)
+                    },
+                    responseType: "json",
+                    signal: controller.signal,
                 },
-                onUploadProgress: (event: AxiosProgressEvent) => {
-                    setLoaded(event.loaded)
-                    setTotal(event.total ?? NaN)
-                },
-                responseType: "json",
-                signal: controller.signal,
-            })
+            )
             ;(async () => {
                 try {
                     const response = await promise
@@ -51,7 +54,7 @@ const UploadProgress: FC<Props> = ({ buffer, filename, onCancel, onComplete, typ
                         throw 500
                     }
                     // :KLUDGE: Make sure it's ready
-                    await new Promise(resolve => setInterval(resolve, 500))
+                    await new Promise(resolve => setTimeout(resolve, 500))
                     onComplete(hash)
                 } catch (e) {
                     if (e instanceof Error) {
@@ -67,23 +70,9 @@ const UploadProgress: FC<Props> = ({ buffer, filename, onCancel, onComplete, typ
             })()
             return () => controller.abort()
         }
-    }, [buffer, contributorUUID, onComplete, token, type])
+    }, [buffer, contributorUUID, existingUUID, onComplete, token, type])
     if (error) {
-        return (
-            <Dialogue>
-                <Speech mode="system">
-                    <p>
-                        <strong>Ack!</strong> There was some kind of error:
-                    </p>
-                    <p>“{String(error)}”</p>
-                </Speech>
-                <UserOptions>
-                    <UserButton icon={ICON_ARROW_LEFT} onClick={onCancel}>
-                        Start over.
-                    </UserButton>
-                </UserOptions>
-            </Dialogue>
-        )
+        return <UploadError error={error} onCancel={onCancel} />
     }
     return (
         <Dialogue>