+++ apps/contribute/src/screens/Uploader/index.tsx
@@ -1,4 +1,4 @@
-import { Hash } from "@phylopic/utils"
+import { Hash, isUUIDv4, UUID } from "@phylopic/utils"
 import { FC, useCallback, useState } from "react"
 import ImageReview from "./ImageReview"
 import { ReviewResult } from "./ImageReview/ReviewResult"
@@ -8,9 +8,9 @@ import UploadProgress from "./UploadProgress"
 export type Props = {
     onCancel: () => void
     onComplete: (hash: Hash) => void
-    value?: Hash
+    existingUUID?: UUID
 }
-const Uploader: FC<Props> = ({ onCancel, onComplete, value }) => {
+const Uploader: FC<Props> = ({ onCancel, onComplete, existingUUID }) => {
     const [fileResult, setFileResult] = useState<FileResult | undefined>()
     const [reviewResult, setReviewResult] = useState<ReviewResult | undefined>()
     const handleImageReviewCancel = useCallback(() => {
@@ -21,7 +21,7 @@ const Uploader: FC<Props> = ({ onCancel, onComplete, value }) => {
         setReviewResult(undefined)
     }, [])
     if (!fileResult) {
-        return <SelectFile onCancel={onCancel} onComplete={setFileResult} value={value} />
+        return <SelectFile onCancel={onCancel} onComplete={setFileResult} isReplacement={isUUIDv4(existingUUID)} />
     }
     if (!reviewResult) {
         return <ImageReview {...fileResult} onCancel={handleImageReviewCancel} onComplete={setReviewResult} />
@@ -29,6 +29,7 @@ const Uploader: FC<Props> = ({ onCancel, onComplete, value }) => {
     return (
         <UploadProgress
             buffer={reviewResult.buffer}
+            existingUUID={existingUUID}
             filename={fileResult?.file.name}
             onCancel={handleUploadProgressCancel}
             onComplete={onComplete}