+++ apps/api-docs/CHANGELOG.md
@@ -23,6 +23,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [2.7.4] - 2026-07-22
+
+### Changed
+
+- Updated version for parity.
+
+## [2.7.3] - 2026-03-17
+
+### Added
+
+- Charset to `Content-Type` headers.
+
+## [2.7.2] - 2026-01-05
+
+### Changed
+
+- Updated version for parity.
+
 ## [2.7.1] - 2025-01-03
 
 ### Added